Chat with YT Video/requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langchain-chroma
langchain-huggingface
chromadb